refactor: move modules auth, income and user from src
