+++ test/domain/auth/controllers/sign-up.controller.e2e.spec.ts
@@ -1,6 +1,6 @@
-import { AuthModule } from "@/domain/auth/auth.module";
-import { SignUpController } from "@/domain/auth/controllers/signup.controller";
-import { SignUpService } from "@/domain/auth/services/signup.service";
+import { AuthModule } from "@/auth/auth.module";
+import { SignUpController } from "@/auth/controllers/signup.controller";
+import { SignUpService } from "@/auth/services/signup.service";
 import { CryptographyModule } from "@/infra/cryptography/cryptography.module";
 import { DatabaseModule } from "@/infra/database/database.module";
 import { INestApplication, ValidationPipe } from "@nestjs/common";
